Use generic decoder and select data based on key
This allows all responses to be handled by the same decoder, and logic that consumes it may now specialize.
This leads to a proper API exception being thrown on invalid token.
